fastapi-app/requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
torch
ultralytics
matplotlib
pillow
tensorflow
pydub
requests